requirements.txt
streamlit
yfinance
pandas
finnhub-python
python-dotenv